[test](be) Split parquet statistics UT coverage
Issue Number: close #xxx

Related PR: #xxx

Problem Summary: Split bloom filter pruning checks out of the end-to-end parquet reader test file into a dedicated parquet statistics UT file, and add a reader-level global row id test that verifies schema exposure and selected-row materialization through the new parquet scan path.

None

- Test: Unit Test
    - python3 build-support/run_clang_format.py --clang-format-executable /opt/homebrew/opt/llvm@16/bin/clang-format -r --style file --inplace false --extensions c,h,C,H,cpp,hpp,cc,hh,c++,h++,cxx,hxx --exclude none be/test/format_v2/parquet/parquet_statistics_test.cpp be/test/format_v2/parquet/parquet_reader_test.cpp
    - ./run-be-ut.sh --run '--filter=ParquetBloomFilterPruningTest.*:NewParquetReaderTest.GlobalRowIdSchemaAndSelectionUseFileRowPosition' -j 8
    - ./build-support/run-clang-tidy.sh --build-dir be/ut_build_ASAN --files be/test/format_v2/parquet/parquet_statistics_test.cpp be/test/format_v2/parquet/parquet_reader_test.cpp
- Behavior changed: No
- Does this need documentation: No
